WL-900 Make the access tool pass through parameters on login
Currently when you try to use access content without a login, all params
in the query string get dropped. It looks like this was fixed for the
portal tool but the behaviour persists in access.

I have moved the logic for constructing the URL to return to to its own
method to tidy things up a bit.

WL-3135 Stop overwriting the login url

The fix for WL-900 removed a null check which resulted in the return url
for a login to be updated for every time the access servlet is hit. This
includes the login pages themselves which is baaad.

The consequence is that after logging in, the user would be redirected to
the login page.