WL-3135 Stop overwriting the login url
The fix for WL-900 removed a null check which resulted in the return url
for a login to be updated for every time the access servlet is hit. This
includes the login pages themselves which is baaad.

The consequence is that after logging in, the user would be redirected to
the login page.